% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/due_date.R
\name{due_date}
\alias{due_date}
\title{Calculate due date}
\usage{
due_date(
  start_date,
  start_type = c("LMP", "conception", "IVF", "fresh_donor", "transfer_day_3",
    "transfer_day_5"),
  cycle = 28
)
}
\arguments{
\item{start_date}{Start date}

\item{start_type}{Start type}

\item{cycle}{Cycle length}
}
\value{
Your due date
}
\description{
Calculate due date
}
\examples{
due_date(lubridate::ymd("2023-02-15"))
}
