% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meds.R
\name{medications-option}
\alias{medications-option}
\alias{set_medications}
\alias{get_medications}
\title{Get or set the pregnancy.medications option}
\usage{
set_medications(medications)

get_medications()
}
\arguments{
\item{medications}{medications}
}
\value{
Both functions invisibly return the current medications setting:
\itemize{
\item get_medications() returns the current setting (a Date object) or NULL if not set
\item set_medications() returns the due date value that was set
}
}
\description{
Functions to get and set the default medications data frame used in the \code{\link[=medications_remaining]{medications_remaining()}} function.
Settings persist for the current R session only, unless added to
.Rprofile. \code{set_medications()} sets the "pregnancy.medications" option and \code{get_medications()} retrieves it.
}
\examples{
# Store original setting
original_medications <- getOption("pregnancy.medications")

# Set the option
set_medications(pregnancy::medications)

# Get the option
get_medications()

# Restore original setting
set_medications(original_medications)

}
\seealso{
\itemize{
\item \code{\link[=medications_remaining]{medications_remaining()}}, \link{medications}
}
}
