% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{medications}
\alias{medications}
\title{An example medications table}
\format{
A data frame with 14 rows and 5 columns:
\describe{
\item{medication}{Name of the medication}
\item{format}{Format of medication}
\item{quantity}{Number taken per day}
\item{start_date}{Date to start taking the medication}
\item{stop_date}{Final date on which the medication is taken. See details.}
}
}
\usage{
medications
}
\description{
A data frame with example medications that might be used during fertility treatment/first trimester.
It is an example of a data frame that might be used as the \code{meds} argument to \code{\link[=meds_remaining]{meds_remaining()}}.
}
\details{
Note that the same medication (prednisolone in this example) has several rows, first because the quantity taken per day changes, then because it needs to be taken on non-consecutive days.
}
\examples{
medications
}
\seealso{
\code{\link[=meds_remaining]{meds_remaining()}}
}
\keyword{datasets}
