% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meds.R
\name{medications_remaining}
\alias{medications_remaining}
\title{Title}
\usage{
medications_remaining(
  medications = NULL,
  group = c("medication", "format"),
  on_date = Sys.Date()
)
}
\arguments{
\item{medications}{medications}

\item{group}{group}

\item{on_date}{date}
}
\value{
prints in console, returns a data frame invisibly
}
\description{
Title
}
\examples{
medications_remaining(on_date = as.Date("2025-04-30"), medications = medications)
}
