% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy-package.R
\docType{package}
\name{pregnancy-package}
\alias{pregnancy}
\alias{pregnancy-package}
\title{pregnancy: Calculate and Track Dates and Medications During Pregnancy}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functionality for calculating pregnancy-related dates and tracking medications during pregnancy and fertility treatment. Calculates due dates from various starting points including last menstrual period and IVF (In Vitro Fertilisation) transfer dates, determines pregnancy progress on any given date, and identifies when specific pregnancy weeks are reached. Includes medication tracking capabilities for individuals undergoing fertility treatment or during pregnancy, allowing users to monitor remaining doses and quantities needed over specified time periods. Designed for those tracking their own pregnancies or supporting partners through the process, making use of options to personalise output messages. For details on due date calculations, see \url{https://www.acog.org/clinical/clinical-guidance/committee-opinion/articles/2017/05/methods-for-estimating-the-due-date}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ellakaye.github.io/pregnancy/}
  \item \url{https://github.com/EllaKaye/pregnancy}
  \item Report bugs at \url{https://github.com/EllaKaye/pregnancy/issues}
}

}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.kaye@gmail.com} (\href{https://orcid.org/0000-0002-7300-3718}{ORCID}) [copyright holder]

}
\keyword{internal}
