% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pregnancy-package.R
\docType{package}
\name{pregnancy-package}
\alias{pregnancy}
\alias{pregnancy-package}
\title{pregnancy: Keep track of dates and medications through pregnancy and fertility treatment}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ella Kaye \email{ella.kaye@gmail.com} (\href{https://orcid.org/0000-0002-7300-3718}{ORCID})

}
\keyword{internal}
